Use the computation stage type in computation stat service
Before this change we were relying on the caller to map the
stage enum to the right long value before sending the request.
This change adds type safety for the stage all the way to the
database write.

Bug: 172104863

+++ src/main/kotlin/org/wfanet/measurement/duchy/service/internal/computationstats/ComputationStatsService.kt
@@ -37,7 +37,7 @@ class ComputationStatsService(
     grpcRequire(metricName.isNotEmpty()) { "Missing Metric name" }
     computationsDatabase.insertComputationStat(
       localId = localComputationId,
-      stage = request.computationStage.toLong(),
+      stage = request.computationStage,
       attempt = request.attempt.toLong(),
       metric = ComputationStatMetric(metricName, request.metricValue)
     )
